% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCDR.R
\name{proj.screen}
\alias{proj.screen}
\title{A function that performs the marginal combined dimension reduction (MCDR)}
\usage{
\method{proj}{screen}(X, Y, A_all, r, screen.by, orth.first = FALSE)
}
\description{
A function that performs the marginal combined dimension reduction (MCDR)
}
