% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{spp.project.lrlda}
\alias{spp.project.lrlda}
\title{class-conditional PCA based on lolR (Eric)}
\usage{
spp.project.lrlda(X, Y, r, xfm = FALSE, xfm.opts = list(), robust = FALSE, ...)
}
\description{
class-conditional PCA based on lolR (Eric)
}
