% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JCDR.R
\name{spp}
\alias{spp}
\title{Joint Combined Dimensionality Reduction (JCDR)}
\usage{
spp(
  X,
  Y,
  r,
  method = "pca+pls",
  orth.first = FALSE,
  second.moment.xfm = FALSE,
  second.moment.xfm.opts = list(),
  first.moment = "delta",
  second.moment = "linear",
  orthogonalize = FALSE,
  robust.first = TRUE,
  robust.second = FALSE,
  ...
)
}
\arguments{
\item{X}{\code{[n, d]} the data with \code{n} samples in \code{d} dimensions.}

\item{Y}{\code{[n]} the labels of the samples with \code{K} unique labels.}

\item{r}{the rank of the projection. Note that \code{r >= K}, and \code{r < d}.}

\item{method}{the method to use for the projection. Defaults to \code{'pca+pls'}. Options include: \code{'mvi'}, \code{'xi'}, \code{'double.proj'}, and \code{'pca+pls'}.}

\item{orth.first}{whether to orthogonalize the projection matrix with first moment and second
moment components. Defaults to \code{FALSE}.}

\item{second.moment.xfm}{whether to use extraneous options in estimation of the second moment component.}

\item{second.moment.xfm.opts}{optional arguments to pass to the \code{second.moment.xfm} option specified.}

\item{first.moment}{the function to capture the first moment. Defaults to \code{'delta'}.
\itemize{
\item{\code{'delta'} capture the first moment with the hyperplane separating the per-class means.}
\item{\code{FALSE} do not capture the first moment.}
}}

\item{second.moment}{the function to capture the second moment. Defaults to \code{'linear'}.}

\item{orthogonalize}{whether to orthogonalize the projection matrix. Defaults to \code{FALSE}.}

\item{robust.first}{whether to use robust PCA for the first moment. Defaults to \code{TRUE}.}

\item{robust.second}{whether to perform PCA on a robust estimate of the second moment component or not.}

\item{...}{additional arguments to pass to the function.}
}
\value{
A list with the following components:
\item{\code{A}}{\code{[d, r]} the projection matrix from \code{d} to \code{r} dimensions.}
\item{\code{centroids}}{\code{[K, r]} the centroids of the classes in the projected space.}
\item{\code{priors}}{\code{[K]} the prior probabilities of the classes.}
\item{\code{ylabs}}{\code{[K]} the unique labels of the classes.}
\item{\code{Xr}}{\code{[n, r]} the data projected into the \code{r} dimensions.}
\item{\code{cr}}{\code{[K, r]} the centroids projected into the \code{r} dimensions.}
\item{\code{second.moment}}{the method used to estimate the second moment.}
\item{\code{first.moment}}{the method used to estimate the first moment.}
}
\description{
A function that performs supervised PCA-PLS.
}
\examples{
library(JCDR)
data(iris)

X <- as.matrix(iris[,1:4])
Y <- iris[,5]
model <- spp(X, Y, r=4)

# use mvi method for projection
model <- spp(X, Y, r=4, method='mvi')
}
