% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spp.utils.info}
\alias{spp.utils.info}
\title{A function that performs basic utilities about the data based on lolR (Eric).}
\usage{
spp.utils.info(X, Y, robust = FALSE, ...)
}
\description{
A function that performs basic utilities about the data based on lolR (Eric).
}
